Machine learning project to predict three major diseases
Health for Future is a web-based application that can predict the three major chronical diseases .It supports the parents and the doctors to get to know about the upcoming threats. This is a user-friendly application with an understandable user interface. The use needs to enter the relevant text fields and the probability will show. This will help the parents to be prepared for the diseases.
I have uploaded the data sets algorithms as well as the python files